+++ changelog.md
@@ -11,3 +11,7 @@
 # 1.0.3
 - Added fallback for no files
 - Prefixed ids with mod id
+
+# 1.0.4
+- Fixed bug where when fading out, it skips back to the first frame
+- Added "Stop video when fading out early" setting